Created and Completed the Mood indicator Project
Completed the documentation and the code of the Project 5: Mood indicator. This project uses a potentiometer as a voltage divider to change the analog input of the voltage. Therefore, we can rotate the servomotor at a desired angle. After some crafting, we added an arrow on top of the servomotor which will let us show the mood that we're currently in. That's why the project is called the Mood indicator.

Further documentation, pictures and improvements to come.